Added eventListeners, functions, and renamed file
Changed name of index file to index.html to keep things clean. I also
finished adding the necessary eventListeners and their respective
functions to create the transitions. The transitions are tied to the
end of the transformation of the click adding/removing open from the
panels. The border will be removed in a later commit and will be kept
purely to see how the text is transitioning the space the text takes.

Fixes-Bug: n/a
Implements-Requirement: n/a