Simplified hover effect
Used psuedo classes from CSS to act as conditionals to prevent the "hover" effect
from going on the open-active panel to make hover stand out more. This hover
only applies to panels that are in the defaul .panel state that get hovered
on. The pointer cursor is also to add to the feel of the hover effect. The
borders are a near blue-tint to constrast with the light colors, while also
applying the triad color theory principle with brown-red being the main color and
blue and yellow being the accent colors.

Fixes-Bug: n/a
Implements-Requirement: n/a