Finalized mouseover and mouseout
Added event listeners to mouseover and mouseout to simulate hovereffect.
This time, there is not bugs with transitionend because the effect removal isnt dependent on the transition
at all but is instead class specific and depends on whether the class is removed or not.
This will replace previous hover effect using psuedoclasses and css :hover

Fixes-Bug: n/a
Implements-Requirement: n/a